requirements.txt
# Production dependencies
requests>=2.31.0
pydantic>=2.5.0
piexif>=1.1.3
Pillow>=10.0.0
imageio-ffmpeg>=0.5.1

# Development dependencies (optional, for local development)
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.1
flake8>=7.0.0
isort>=5.13.0
